fix: Add extraline in private ssh key (#100)
The git-clone task requires the usage of a private ssh-key to clone the
repository, and it is imperative that the key is in the correct format.
Specifically, a new line at the end of the private key file is mandatory.
This patch addresses the issue by adding the necessary newline.

Additionally, we are ok with the idea of appending a new line to an
existing one since this is also a correct format and may be easier to
implement.


Jira: EPMDEDP-13125

Related: https://github.com/epam/edp-tekton/issues/100

Change-Id: I7bf4f1ca67df420e94feba08cf6f5b5941baed3b