fix: Ensure build pipeline is triggered only for Merge Events (#128)
Updated the GitHub EventListener to check for merge events
before triggering the build pipeline for closed pull requests.

This fixes the issue where the build pipeline was being
triggered for all closed pull requests, regardless of whether
they were merged or not. Now, the build pipeline will only
be triggered when a pull request is closed due to a merge event.

JIRA: EPMDEDP-13280

Signed-off-by: Sergiy Kulanov <sergiy_kulanov@epam.com>
Change-Id: I6dfd8e9435788d126379f7709e4ec6be38c717bc