feat: Enhance Pipeline Triggering Process (#263)
We are refining the pipeline triggering process to
allow specific pipelines to be defined for code branches
directly within the codebasebranch API spec.

This enhancement addresses the previous limitation
of dynamic pipeline naming and enables the direct
retrieval of pipeline names from codebasebranch for processing.

The proposed changes include updating the TriggerTemplate, TriggerBinding

Signed-off-by: Sergiy Kulanov <sergiy_kulanov@epam.com>